Added optional per-game configuration file support.
Uses a complimentary `$(romname).sam` file next to game rom to specify custom properties.
Currently allows overriding the countdown value per game. E.G.

Legend of Zelda - A Link to the Past (USA).sfc.sam
```
counter=172
```

This configuration gives the whole opening sequence a single playthrough for
`A Link to the Past`, but will return to the default countdown value on
subsequent games.