Add Atari 2600 5200 7800 and Lynx
Atari2600 was, long ago, merged with the ATARI7800 core, and, the rom files need to be in the ATARI7800 folder for mgl to work, until something is changed in the cores/mgl code, you can use the pathsextra fields to separate them into separate directories, under ATARI7800 if you wish, which will keep the gamelist creator from scanning both sets of files for each core.

usedefaultpaths only takes a couple of seconds, so, is not "slow" and will be enabled by default. You can keep it set to No or, set it to No if you'd like, and, it will continue to work, somewhat, but, will become unsupported.

Everdrive mappings have been made obsolete, and have been removed, this affects nobody, as we don't use them for anything.