Fix no core repeat
Cores will now randomly go through the entire list, and play one game from each one, before repeating a core.

Fixed gamelist creation to always create all game lists (as long as the options are correctly set) when regenerating lists, regardless of what cores you have selected in the ini.

Fix a long standing, invisible problem with the skip command that will be the basis of a rewrite of the commandline and internal command parsing system.