Fix Arcade Gamelist creation
If you don't specify "/_Organized" or some directory beneath it, that whole directory structure will be ignored, since all the mra files in them are symlinks to other mras and can cause duplicate games to be added to the list

However, if you do set your arcadepathextra setting in the ini to any directory _Organize or below,  it won't ignore them.

Also tidied up a few incidentals while working on this.