Major update
Mainly finishes integrating the load core functionality. Also, Arcade and Amiga can now be played when playcurrentgame=yes There may still be some fixing and tweaking to be done to all of this, but, I need to push the changes, as they are, so more people can hopefully help find where things need improvement.

This also gets rid of the useless main function and replaces it with what used to be the parse_cmd function.