Fix issue with config files being the wrong case
Hopefully, the cores can be fixed so they don't create lowercase file names, but, this will make sure that cores that need their config files, such as the Amiga core, and, can't find them, because they don't match the actual core name won't prevent the core from working, properly.

The reason this is needed is we bind mount a /tmp directory to /media/fat/config so we can mute the volume, without muting these cores when you play them outside of SAM, especially if you have playcurrengame set to yes the tmpfs file system is case sensitive, but, the config directory is located on the SD card, which is FAT32, and, not case sensitive, so, some devs aren't taking care to make sure they are creating the config files with the correct case.