Add support for atari5200 skipmessage per game
This can also be used for adding other cores that have only some games that require a button press to bypass start screens and start the attract mode

Atari5200 has some games that ask you to press right or left, to choose if it is a single chip or dual chip cartridge, this automates this process. If you find a game that this doesn't work for, post the name of the game on the issue tracker, and we will add it.