Migrate more to the system.ini
add md5sum for the ini, so the values will be loaded on start, if the ini  file has changed.

add reset command that will delete the temporary settings files
Generate CORE_PATH_FINAL and pass it as defaultpaths, although, if you disable default paths, it should still pass the ini settings along.