Compatibility with the main MiSTer project from the 'aitorgomez' fork.
A new variable named configpath is set. By default, it points to /media/fat/config. However, if the cfgcore_subfolder entry is defined in MiSTer.ini, it is used to extend the path. Likewise, if cfgarcade_subfolder is set, its configuration folder will be used in the arcade cores

More information about the aitorgomez fork can be found here:

https://github.com/spark2k06/Main_MiSTer